refactor(web): 認証ハンドラの共通化をスライス内完結の設計に改める
server/error/ ディレクトリと横断ヘルパ active-user-login.ts を撤去し、
スライスを跨ぐ共有を throw-http-error.ts のみに戻す。

- unwrapOrThrowHttp は throw-http-error.ts へ統合し、error/ ディレクトリを廃止する
- パスワードログインとパスキー認証で共有していた respondActiveUserLogin は、
  ログイン応答の責務がスライスを跨いで曖昧になるためやめる
- 代わりに passkey スライス既存の createPasskeyActionHandler を検証付き
  Context に対応させ、authentication-verify をスライス内ファクトリで完結させる
- createSession は sessions スライス内で自己完結させる

この形でも similarity しきい値 0.9 で重複 0 件を維持する
（旧ハブの残存類似は最大 75.38%）。各ハンドラの挙動は不変。
