+++ application/apps/web/src/server/passkey/error.ts
@@ -1,5 +1,5 @@
 import { PasskeyRegistrationError, PasskeyVerificationError } from '@trend-diary/authentication'
-import throwHttpErrorByTable, { type ErrorStatusTable } from '@/server/error/throw-http-error'
+import throwHttpErrorByTable, { type ErrorStatusTable } from '@/server/throw-http-error'
 
 // 認証後の active_user 解決失敗（ActiveUserNotFoundError 等）はサーバ側の不整合なので、
 // ここでは写像せず default の 500 へ倒す。account のエラー写像は passkey ハンドラから分離する
