+++ application/apps/web/src/server/registrations/error.ts
@@ -1,5 +1,5 @@
 import { UserAlreadyExistsError } from '@trend-diary/authentication'
-import throwHttpErrorByTable, { type ErrorStatusTable } from '@/server/error/throw-http-error'
+import throwHttpErrorByTable, { type ErrorStatusTable } from '@/server/throw-http-error'
 
 const ERROR_STATUS_TABLE: ErrorStatusTable = [[UserAlreadyExistsError, 409]]
