+++ application/apps/web/src/server/sessions/handler/create.ts
@@ -1,10 +1,11 @@
 import { authClientConfig, PasswordAuthClient } from '@trend-diary/authentication'
-import { authInputSchema } from '@trend-diary/domain/account'
+import getRdbClient from '@trend-diary/datastore/rdb'
+import { authInputSchema, createAccountUseCase } from '@trend-diary/domain/account'
+import CONTEXT_KEY from '@/middleware/context'
 import zodValidator, { type ZodValidatedContext } from '@/middleware/zod-validator'
-import respondActiveUserLogin from '@/server/active-user-login'
 import { assertCaptchaVerified } from '@/server/captcha'
-import unwrapOrThrowHttp from '@/server/error/unwrap-or-throw-http'
 import throwHttpError from '@/server/sessions/error'
+import { unwrapOrThrowHttp } from '@/server/throw-http-error'
 
 export const authInputValidator = zodValidator('json', authInputSchema)
 
@@ -19,5 +20,14 @@ export default async function createSession(c: ZodValidatedContext<[typeof authI
     throwHttpError,
   )
 
-  return respondActiveUserLogin(c, user.id, throwHttpError, 'session created')
+  const accountUseCase = createAccountUseCase(getRdbClient(c.env.DB))
+  // 認証成功後に active_user が無いのは孤児 auth ユーザー等のサーバ不整合なので、404 ではなく 500 に倒す
+  const activeUser = unwrapOrThrowHttp(
+    await accountUseCase.resolveActiveUser(user.id),
+    throwHttpError,
+  )
+
+  c.get(CONTEXT_KEY.APP_LOG).info('session created', { activeUserId: activeUser.activeUserId })
+
+  return c.json({ displayName: activeUser.displayName }, 200)
 }
